Refine date use-case comment
The date property has an ambiguous use-case because it can
be either a dateCreated value or a dateEdited value. It depends
entirely on the use-case.

Changelogs for example, should only use the date as a dateCreated
value because even if the changelog gets updated, the content is only
applicable to that date, but a FAQ section would use the date as a
dateEdited value which lets users know what information is most up
to date.