Add markdown rendering
Because these pages were being parsed on the users machine,
this incurred a huge dependency overhead. I realized there's
no reason to add that dependency if we can render the
markdown before the client machine ever sees it.

The default action is still to preserve markdown, for those
use-cases where it makes sense. But in my current use-case
where all pages are static, it makes no sense to burden the
client machine with a worthless dependency.