Add URI to pages
This allows there to be a standardized way to navigate to
pages without having to pre-process it. In most cases, doing
what I did here is negligable performance-wise, but when
prototyping you don't really want to have to worry about
how to linkify your pages.

This is a convenience method at the very least.