Rename isTesting to isLogging and change string requirement
Since this package is going to be published to NPM, it will
now be used as more than just a standalone app. If it needs
to integrate with other logic, then making an 'is-testing'
assumption no longer makes sense.

It makes much more sense to apply a silent flag, so that
the parent app can utilize its own output.