Cannibalize markdown-it-video for personal use
The package had not been pushed to NPM in several years, so I
had no choice but to copy the master code and adapt it myself.

I decided to strip it down to only the parts I needed as well, so
all in all this gives us a much slimer implamentation that actually
works.